Move CTransaction from main.h to core.h.
